Update `nightly` toolchain in macro crate
