+++ macro/Cargo.toml
@@ -28,4 +28,4 @@ version-sync = "0.9.4"
 default = []
 # Uses custom test frameworks APIs together with a generous spicing of hacks
 # to include arguments in the names of the generated tests.
-nightly = []
+nightly = ["proc-macro2/span-locations"]
